# Only tessellate joins between curve segments.