Instead of using a flattening iterator, implement the `PathBuilder` trait for `StrokeBuilder` and use `flattened_for_each`. Only tessellate special line joins between curves instead of for each flattened segment.

This prevents the tessellator from inserting joins in the middle of curve segments so it is intended as a correctness fix, but it also incidentally improved performance, even for miter joins (probably because `flattened_for_each` implementations are faster than the flattening iterators).

Before:

```
test stroke_logo_bevel                   ... bench:   6,964,774 ns/iter (+/- 750,889)
test stroke_logo_miter                   ... bench:   6,292,831 ns/iter (+/- 702,244)
test stroke_logo_round                   ... bench:   8,724,966 ns/iter (+/- 943,608)
```

After:

```
test stroke_logo_bevel                   ... bench:   5,820,940 ns/iter (+/- 169,468)
test stroke_logo_miter                   ... bench:   5,488,707 ns/iter (+/- 56,132)
test stroke_logo_round                   ... bench:   6,968,883 ns/iter (+/- 779,524)
```